extra lazy loading (#139)
* introduce LazyArrayDict

* update lazy loading + unitests

* update dandi_utils and minor improvements

* update perich miller

* refractor experiment names

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* update cfg

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* fix sampler

* fix sampler

* rm getattr trigger when printing representation

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* minor fixes

* add extra extra lazy

* add extra extra lazy

* add extra extra lazy

* add extra extra lazy

* add extra extra lazy

* add extra extra lazy

* Interval Arithmetic (#136)

* Allows `track_batch`/`mask`/`mask8` to work on lists on objects (#134)

* Update jsonschema to new version in requirements.txt (#135)

A newer version is required by jupyter-lab. I didn't see any fails from
this upgrade.

* Adds & and | operations for Intervals + tests

* update doc

* update test_data

* update core

* Fixes for `extra_lazy_loading` (#140)

* Domain should be a required parameter

* Changes to get prepare_data on mc_maze_small working

* Removes accidental breakpoint()

* Interval(start,end) -> Interval.from_list() in prep data

* data.slice() updates domain

* Fixes LazyIrregTS slicing

* Adds Data.slice_start_time

* Simple test_data fix

* Renames slice_start_time to slice_start + adds to test

* Better error reporting in data.py

* Doc fix

* slice_start -> absolute_start

* update prepare_data

* update prepare_data

* update validation

* fix import

* fix import

* fix import

* fix import

* fix import

* fix decoder registry

* fix tokenizer

* fix tokenizer

* fix tokenizer

* move to device

* move to device

* move to device

* validation

* validation

* validation

* validation

* validation

* validation

* validation

* validation

* validation

* validation

* validation

* update weight_registry

* Memory Efficient Attention + Chaining (#141)

* Domain should be a required parameter

* Changes to get prepare_data on mc_maze_small working

* Removes accidental breakpoint()

* Interval(start,end) -> Interval.from_list() in prep data

* data.slice() updates domain

* Fixes LazyIrregTS slicing

* Adds Data.slice_start_time

* Simple test_data fix

* Renames slice_start_time to slice_start + adds to test

* Better error reporting in data.py

* Doc fix

* slice_start -> absolute_start

* Refactors Rotary*Attention for xformers

* attn_mask -> context_mask

* Modify perceiver_rotary to work with chained inputs

* WIP: Memory efficient. Working default attention

* Working memory_efficient_attention!

* Memory efficient bugfix

* Cleanup

* add dilate and difference

---------

+++ kirby/data/sampler.py
@@ -76,7 +76,7 @@ class RandomFixedWindowSampler(torch.utils.data.Sampler):
                         DatasetIndex(
                             session_name, t.item(), (t + self.window_length).item()
                         )
-                        for t in torch.arange(start + left_offset, end, self.window_length)
+                        for t in torch.arange(start + left_offset, end, self.window_length, dtype=torch.float64)
                         if t + self.window_length <= end
                     ]
                 
@@ -154,7 +154,7 @@ class SequentialFixedWindowSampler(torch.utils.data.Sampler):
                         DatasetIndex(
                             session_name, t.item(), (t + self.window_length).item()
                         )
-                        for t in torch.arange(start, end, self.step)
+                        for t in torch.arange(start, end, self.step, dtype=torch.float64)
                         if t + self.window_length <= end
                 ]
